refactor: deduplicate config types
